Fix clicking button container bug
Clicking the button container after filling out the equation led to a
bug with text input from the container being read in and displayed. By
adding the additional boolean requirement of the input being an operator
to the last if statement that was causing the bug, the bug no longer
exists.